+++ nextjs-dashboard/app/ui/search/types.ts
